+++ SpotifyAPI.Web/Models/Response/FullArtist.cs
@@ -11,7 +11,7 @@ namespace SpotifyAPI.Web
     public string Id { get; set; } = default!;
     public List<Image> Images { get; set; } = default!;
     public string Name { get; set; } = default!;
-    public int Popularity { get; set; }
+    public double Popularity { get; set; }
     public string Type { get; set; } = default!;
     public string Uri { get; set; } = default!;
   }
